feat: 예약된 테마 삭제 시 예외 처리 로직 추가

+++ src/main/java/roomescape/service/ThemeService.java
@@ -45,6 +45,9 @@ public class ThemeService {
         if (!themeRepository.existsById(id)) {
             throw new NoSuchElementException("[ERROR] 존재하지 않는 ID입니다.");
         }
+        if (reservationRepository.existsByThemeId(id)) {
+            throw new IllegalArgumentException("[ERROR] 해당 테마의 예약이 존재합니다.");
+        }
         themeRepository.delete(id);
     }
